Temporarily move all translations into qcStepNameTranslations
